% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpsschk.R
\name{dpsschk}
\alias{dpsschk}
\title{function to calculate tapers}
\usage{
dpsschk(N, k, nw, Fs)
}
\arguments{
\item{N}{(number of samples}

\item{Fs}{sampling frequency}

\item{K}{- number of tapers, often 2*nw for  spectrum estimation purposes.}

\item{NW}{- A positive double-precision number, the time-bandwidth parameter}
}
\value{
tapers
}
\description{
Called by spectral estimation routines to generate the tapers
}
