% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfgrid.R
\name{getfgrid}
\alias{getfgrid}
\title{function that gets the frequency grid associated with a given fft based computation}
\usage{
getfgrid(Fs, nfft, fpass)
}
\arguments{
\item{Fs}{sampling frequency associated with the data}

\item{nfft}{number of points in fft}

\item{fpass}{band of frequencies at which the fft is being calculated (fmin fmax) in Hz}
}
\value{
f         frequencies)

findx     index of the frequencies in the full frequency grid
}
\description{
Called by spectral estimation routines to generate the frequency axes
}
