% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp.event.times.R
\name{hpp.event.times}
\alias{hpp.event.times}
\title{simulate events from a homogeneous Poisson process}
\usage{
hpp.event.times(rate, num.events, t0 = 0, tT = 1, sig = 4)
}
\arguments{
\item{num.events}{}
}
\value{
aa1 vector of event times
}
\description{
simulate events from a homogeneous Poisson process
}
