% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{integrationSites}
\alias{integrationSites}
\title{Two samples of viral integration sites.}
\format{
A list with two data frames
\describe{
  \item{London}{Distance along the genome (combined chromosomes) measued in basepairs. There are 72775 integration sites}
  \item{Paris}{Distance along the genome (combined chromosomes) measued in basepairs. There are 308538  integration sites}
  ...
}
}
\source{
Supplied by the authors
}
\usage{
integrationSites
}
\description{
X-linked severe combined immunodeficiency (SCID-X1) is an inherited
disease caused by inactivating mutations in the gene
encoding the interleukin 2 receptor common gamma chain (IL2RG), which is located on the X-chromosome.
SCID-X1 gene therapy clinical trials were conducted in Paris and London. The clinical results differed and the
question was raised of the effect of different transduction conditions on transduced cells.  Human peripheral blood (PB)
CD34+ cells were transduced with $\gamma$-retroviral vectors according to the transduction protocols employed in the
initial clinical trials.  14 Cells cultured under London trial conditions showed relatively higher retention of CD34
expression at the completion of the transduction period, compared to a loss of CD34-positivity in about half the cells
transduced under the Paris trial conditions. Paris conditions, on the other hand, promoted higher levels of both
proliferation and transduction, see \cite{Hallwirth.et.al.2015}.
}
\keyword{datasets}
