% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtfftpt.R
\name{mtfftpt}
\alias{mtfftpt}
\title{Multi-taper fourier transform for point process given as times}
\usage{
mtfftpt(data, tapers, nfft, t, f, findx)
}
\arguments{
\item{tapers}{from dpss}

\item{nfft}{length of padded data}

\item{t}{time points at which tapers are calculated}

\item{f}{frequencies at which fft is caculated}

\item{findx}{index corresponding to frequencies f}
}
\value{
J fft in form frequency index x taper index

Msp (number of spikes per sample in each channel

Nsp (number of spikes in each channel
}
\description{
Multi-taper fourier transform for point process given as times
}
