% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtspectrumpt.R
\name{mtspectrumpt}
\alias{mtspectrumpt}
\title{multi-taperd spectrum of data preseted as spike locations}
\usage{
mtspectrumpt(
  PP,
  Fs = 3000,
  err = c(2, 0.05),
  fpass = c(0, 1500),
  pad = 0,
  nw = 50,
  k = 2 * nw - 1,
  fscorr = 0
)
}
\arguments{
\item{PP}{the data}

\item{Fs}{sampling frequency}

\item{err}{sampling frequency}

\item{fpass}{sampling frequency}

\item{pad}{sampling frequency}

\item{fscorr}{sampling frequency}

\item{tapers}{sampling frequency}
}
\value{
S
}
\description{
multi-taperd spectrum of data preseted as spike locations
}
