% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherencypt.R
\name{coherencypt}
\alias{coherencypt}
\title{Multi-taper coherency - point process times}
\usage{
coherencypt(
  data1,
  data2,
  Fs = 3000,
  fpass = c(0, 1500),
  pad = 0,
  nw = 50,
  k = 2 * nw - 1
)
}
\arguments{
\item{data1}{array of spike times}

\item{data2}{array of spike times}

\item{Fs}{sampling frequency}

\item{fpass}{(fmin, fmax), frequency band to be used in the calculation. Defaults to (0, Fs/2)}

\item{pad}{(padding factor for the FFT). Not currently used, i.e. no padding is done}

\item{nw}{A positive double-precision number, the time-bandwidth paramter for the tapers}

\item{k}{A positive integer, the number of tapers, defaults to  2*nw-1}

\item{fscorr}{finite size corrections -- not implemented yet}
}
\value{
S1     the estimated spectrum for data1

S2     the estimated spectrum for data2

C12    the coherency

f      the frequencies
}
\description{
Multi-taper coherency - point process times
}
