% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherencypt.R
\name{coherencypt}
\alias{coherencypt}
\title{Multi-taper coherency - point process times}
\usage{
coherencypt(
  data1,
  data2,
  Fs = 3000,
  err = c(2, 0.05),
  fpass = c(0, 1500),
  pad = 0,
  nw = 50,
  k = 2 * nw - 1,
  fscorr = 0
)
}
\arguments{
\item{data1}{array of spike times}

\item{data2}{array of spike times}

\item{fscorr}{finite size corrections}

\item{params}{}

\item{t}{time grid over which the tapers are to be calculated

 @return  C          frequencies)
 @return findx     index of the frequencies in the full frequency grid}
}
\description{
Multi-taper coherency - point process times
}
