% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpsschk.R
\name{dpsschk}
\alias{dpsschk}
\title{function to calculate tapers}
\usage{
dpsschk(tapers, N, Fs)
}
\arguments{
\item{tapers}{NW K - time-bandwidth product, number of tapers}

\item{N}{(number of samples}

\item{Fs}{sampling frequency}
}
\value{
tapers
}
\description{
Called by spectral estimation routines to generate the tapers
}
