% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtspectrumpt.R
\name{mtspectrumpt}
\alias{mtspectrumpt}
\title{multi-tapered spectrum of data presented as spike locations}
\usage{
mtspectrumpt(
  PP,
  Fs = 3000,
  fpass = c(0, Fs/2),
  pad = 0,
  nw = 50,
  k = 2 * nw - 1,
  fscorr = 0
)
}
\arguments{
\item{PP}{the data}

\item{Fs}{sampling frequency}

\item{fpass}{(fmin, fmax), frequency band to be used in the calculation. Defaults to (0, Fs/2)}

\item{pad}{(padding factor for the FFT). Not currently used, i.e. no padding is done}

\item{nw}{A positive double-precision number, the time-bandwidth paramter for the tapers}

\item{k}{A positive integer, the number of tapers, defaults to  2*nw-1}
}
\value{
S  the estimated spectrum

f  the frequencies
}
\description{
multi-tapered spectrum of data presented as spike locations
}
